pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sleeplab-extractor"
description = "Tools for extracting and preprocessing data from sleeplab format to various studies' specific needs."
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
  { name = "Riku Huttunen", email = "riku.huttunen@uef.fi" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "numpy",
  "pyyaml",
  "scipy",
  "sleeplab-format @ git+https://github.com/UEF-SmartSleepLab/sleeplab-format@main"
]
dynamic = ["version"]

[project.scripts]
sleeplab-extract = "sleeplab_extractor.extract:run_cli"

[project.urls]
Documentation = "https://github.com/UEF-SmartSleepLab/sleeplab-extractor#readme"
Issues = "https://github.com/UEF-SmartSleepLab/sleeplab-extractor/issues"
Source = "https://github.com/UEF-SmartSleepLab/sleeplab-extractor"

[tool.hatch.version]
path = "sleeplab_extractor/version.py"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=sleeplab_extractor --cov=tests {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "sleeplab_extractor/version.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.hatch.metadata]
allow-direct-references = true
